requirements.txt
tensorflow>=2.14.0
keras>=3.0.0
numpy
pandas
scikit-learn
matplotlib
seaborn
streamlit
Pillow
opencv-python-headless
